+++ src/Pos/Webhook/WebhookController.php
@@ -11,6 +11,7 @@ use Psr\Log\LoggerInterface;
 use Shopware\Core\Framework\Context;
 use Shopware\Core\Framework\DataAbstractionLayer\EntityRepositoryInterface;
 use Shopware\Core\Framework\DataAbstractionLayer\Search\Criteria;
+use Shopware\Core\Framework\Routing\Annotation\Acl;
 use Shopware\Core\Framework\Routing\Annotation\RouteScope;
 use Shopware\Core\System\SalesChannel\SalesChannelEntity;
 use Swag\PayPal\Pos\Api\Webhook\Webhook;
@@ -65,6 +66,7 @@ class WebhookController extends AbstractController
      *     name="api.action.paypal.pos.webhook.registration.register",
      *     methods={"POST"}
      * )
+     * @Acl({"sales_channel.editor"})
      */
     public function registerWebhook(string $salesChannelId, Context $context): Response
     {
@@ -79,6 +81,7 @@ class WebhookController extends AbstractController
      *     name="api.action.paypal.pos.webhook.registration.unregister",
      *     methods={"DELETE"}
      * )
+     * @Acl({"sales_channel.deleter"})
      */
     public function unregisterWebhook(string $salesChannelId, Context $context): Response
     {