+++ src/Resources/app/administration/src/module/swag-paypal-pos/extension/sw-sales-channel-menu/index.js
@@ -3,7 +3,6 @@ import { PAYPAL_POS_SALES_CHANNEL_TYPE_ID } from '../../../../constant/swag-payp
 const { Component } = Shopware;
 
 Component.override('sw-sales-channel-menu', {
-
     methods: {
         createMenuTree() {
             this.$super('createMenuTree');