+++ src/Resources/app/administration/src/module/swag-paypal-pos/page/swag-paypal-pos-detail-runs/index.js
@@ -9,7 +9,8 @@ Component.register('swag-paypal-pos-detail-runs', {
 
     inject: [
         'SwagPayPalPosApiService',
-        'repositoryFactory'
+        'repositoryFactory',
+        'acl'
     ],
 
     mixins: [