+++ src/Resources/app/administration/src/module/swag-paypal/components/swag-paypal-behavior/index.js
@@ -7,6 +7,10 @@ const { Component } = Shopware;
 Component.register('swag-paypal-behavior', {
     template,
 
+    inject: [
+        'acl'
+    ],
+
     props: {
         actualConfigData: {
             type: Object,